Do not try to delete dirvish#open
I am seeing the following error due to this in both Vim and Neovim:

> E131: Cannot delete function dirvish#open: It is in use

Stepping into the call to it before it loads the autoload file, calls
the function and returns from it.  Therefore it is not clear to me
how/why it is in use, but I also cannot see why it should get deleted.

This statement is there since the beginning (cd94a24).